Add files via upload
The dataset contains a public dataset downloaded from Kaggle.
Dataset Description: The data is used as the basis of building a model from a direct marketing campaign of the banking institution. The marketing campaign data records were based on phone calls to assess if a specific product (bank term deposit) would be  "yes" or "no" subscribed.